PHD can use differenet Kalman filter algorithms.
Here multiple changes have been made. Firstly the PHD algorithm was
rewritten as class to streamline the interface and reduce number of
inputs regularly necessary for algorithm to perfrom.

Secondly Kalman filter algorithm which needs to be used is passed in
as an object during initialization. Only assumption here is that
passed in object is subclass of Kalman ABC from estimation package.

Lastly the function was rewritten so that weights states and covariances
are internal states.

This closes #2 issue