Restructure the builders for Kalman filters
The Kalman filters and their builder classes are logically connecnted.
Because of this a large refactoring has been undertaken to place them into a
single Python module. In this new structure the builders are inner classes
of respective filters and perform the necessary checks to ensure that all
parameters of the filter are properly instantiated.