Correct the implementation of GaussianMixtureMixin reduce method
The reduce() method of GaussianMixtureMixin had multiple errors which led to
false computations of reduced mixtures. The following errors were present:

- The Mahalanobis distances were not computed correctly
- The selection process for components to be merged was not correct
- The computation of new covariances was incorrect
- The removal of components already merged from list of remaining components
  was not correct.

First error was resolved by correcting the computation. The mahalanobs distance
is the square root of quadratic form of difference of two vectors over the
inverse of covariance matrix. This quadratic form can be computed for many
vectors at once, by placing all of the input vectors in matrices, and then
extracting the diagonal of the resulting matrix. As only diagonal is necessary
this can be made significantly faster by computing only this diagonal entry.

The selection process was improved by making all of the selections at single
place in code, thus providing better code visibility.

The computation of new covariances was false because the input difference was
computed as difference between the new state and means, while the correct
procedure computes the difference between the new state and selected states.

Last correction was made with respect to removal of merged components from the
list of remaining components to be merged. The np.delete method did not work in
some special cases, so I switched to more robust selection notation.